pyproject.toml
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''

[tool.poetry]
name = "aws_iot_sdk_sample"
version = "0.1.0"
description = "A sample showcasing basic usage of AWS IoT Core"
authors = ["Andrei Olar <andrei.olar@samlex.ro>"]

[tool.poetry.dependencies]
python = "^3.10"
awsiotsdk = "^1.11.1"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
